[ref] Moves markedOptions
