+++ package.json
@@ -39,6 +39,7 @@
     "core-js": "^2.5.7",
     "moment": "^2.22.2",
     "ngx-markdown": "^6.2.0",
+    "marked": "^0.5.0",
     "prismjs-loader": "^0.0.4",
     "rxjs": "^6.2.1",
     "rxjs-compat": "^6.2.1",
